resume_ranker_ai/requirements.txt
streamlit>=1.15.0
spacy>=3.4.0
PyPDF2>=2.10.0
python-docx>=0.8.11
pdfplumber>=0.7.0
scikit-learn>=1.0.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.4.0/en_core_web_md-3.4.0-py3-none-any.whl